# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Web3 backend for Sukuk (Islamic bonds) on Base Testnet, providing APIs for dApps and frontends while processing blockchain events from an indexer.",
        "title": "Sukuk POC Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.sukuk-poc.com/support",
            "email": "support@sukuk-poc.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/sukuks": {
            "get": {
                "description": "Get a list of all Sukuk series with optional filtering by company and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sukuk Series"
                ],
                "summary": "List all Sukuk series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID to filter by",
                        "name": "company_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status to filter by (active, paused, matured)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Sukuk series",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.SukukSeriesListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sukuks/{id}": {
            "get": {
                "description": "Get detailed information about a specific Sukuk series including company, investments, yield claims, and redemptions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sukuk Series"
                ],
                "summary": "Get Sukuk series details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sukuk Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukuk series details",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.SukukSeriesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Sukuk series not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sukuks/{id}/holders": {
            "get": {
                "description": "Get a list of current active investors holding the specified Sukuk series",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sukuk Series"
                ],
                "summary": "Get Sukuk holders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sukuk Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of active investments",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.SukukHoldersResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/yield-claims": {
            "get": {
                "description": "Get a list of all yield claims with optional filtering by status and investor address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yield Claims"
                ],
                "summary": "List all yield claims",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Yield claim status to filter by (pending, claimed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Investor wallet address to filter by",
                        "name": "investor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of yield claims",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.YieldClaimListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/yield-claims/investor/{address}": {
            "get": {
                "description": "Get all yield claims made by a specific wallet address with optional status filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yield Claims"
                ],
                "summary": "Get yield claims by investor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Investor wallet address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Yield claim status to filter by (pending, claimed)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Investor's yield claims",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.YieldClaimListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid address",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/yield-claims/sukuk/{sukukId}": {
            "get": {
                "description": "Get all yield claims for a specific Sukuk series with optional status filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yield Claims"
                ],
                "summary": "Get yield claims by Sukuk series",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sukuk Series ID",
                        "name": "sukukId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Yield claim status to filter by (pending, claimed)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukuk series yield claims",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.YieldClaimListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/yield-claims/{id}": {
            "get": {
                "description": "Get detailed information about a specific yield claim including Sukuk series, company, and investment data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yield Claims"
                ],
                "summary": "Get yield claim details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Yield Claim ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yield claim details",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.YieldClaimResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Yield claim not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_kadzu_sukuk-poc-be_internal_models.Investment": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Investment amount in IDRX",
                    "type": "string",
                    "example": "1000000000000000000000"
                },
                "block_number": {
                    "type": "integer",
                    "example": 12345678
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "investment_date": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "investor_address": {
                    "type": "string",
                    "example": "0x1234567890123456789012345678901234567890"
                },
                "investor_email": {
                    "type": "string",
                    "example": "investor@example.com"
                },
                "redemptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Redemption"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "sukuk_series": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries"
                        }
                    ]
                },
                "sukuk_series_id": {
                    "type": "integer",
                    "example": 1
                },
                "token_amount": {
                    "description": "Equivalent Sukuk tokens received",
                    "type": "string",
                    "example": "1000000000000000000000"
                },
                "transaction_hash": {
                    "type": "string",
                    "example": "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "yield_claims": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.YieldClaim"
                    }
                }
            }
        },
...
        "github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries": {
            "type": "object",
            "properties": {
                "company": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Company"
                        }
                    ]
                },
                "company_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Sustainable infrastructure financing sukuk"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "investments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Investment"
                    }
                },
                "is_redeemable": {
                    "description": "Can investors redeem early",
                    "type": "boolean",
                    "example": true
                },
                "legal_documents": {
                    "description": "JSON array of document paths",
                    "type": "string",
                    "example": "[\"legal1.pdf\", \"legal2.pdf\"]"
                },
                "maturity_date": {
                    "type": "string",
                    "example": "2027-12-31T00:00:00Z"
                },
                "max_investment": {
                    "description": "Maximum investment amount (optional)",
                    "type": "string",
                    "example": "100000000000000000000"
                },
                "min_investment": {
                    "description": "Minimum investment amount",
                    "type": "string",
                    "example": "1000000000000000000"
                },
                "name": {
                    "type": "string",
                    "example": "Green Sukuk Series A"
                },
                "outstanding_supply": {
                    "description": "Currently issued",
                    "type": "string",
                    "example": "500000000000000000000000"
                },
                "payment_frequency": {
                    "description": "Payments per year (quarterly = 4)",
                    "type": "integer",
                    "example": 4
                },
                "prospectus": {
                    "description": "PDF file path",
                    "type": "string",
                    "example": "/uploads/prospectus/sukuk_1.pdf"
                },
                "redemptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Redemption"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "symbol": {
                    "type": "string",
                    "example": "GSA"
                },
                "token_address": {
                    "type": "string",
                    "example": "0xabcdefabcdefabcdefabcdefabcdefabcdefabcd"
                },
                "total_supply": {
                    "description": "Use string for big numbers",
                    "type": "string",
                    "example": "1000000000000000000000000"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "yield_claims": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.YieldClaim"
                    }
                },
                "yield_rate": {
                    "description": "Annual yield rate (e.g., 0.0850 for 8.5%)",
                    "type": "number",
                    "example": 0.085
                }
            }
        },
        "github_com_kadzu_sukuk-poc-be_internal_models.YieldClaim": {
            "type": "object",
            "properties": {
                "block_number": {
                    "description": "Set when claimed",
                    "type": "integer"
                },
                "claimed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "description": "Yield claims can expire",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "investment": {
                    "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Investment"
                },
                "investment_id": {
                    "type": "integer"
                },
                "investor_address": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.YieldClaimStatus"
                },
                "sukuk_series": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries"
                        }
                    ]
                },
                "sukuk_series_id": {
                    "type": "integer"
                },
                "transaction_hash": {
                    "description": "Set when claimed",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "yield_amount": {
                    "description": "Yield amount in IDRX",
                    "type": "string"
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request data"
                }
            }
        },
...
        "internal_handlers.MetaInfo": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "internal_handlers.SukukHoldersResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 150
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.Investment"
                    }
                }
            }
        },
...
        "internal_handlers.SukukSeriesListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 10
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/internal_handlers.MetaInfo"
                }
            }
        },
        "internal_handlers.SukukSeriesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.SukukSeries"
                }
            }
        },
...
        "internal_handlers.YieldClaimListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 50
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.YieldClaim"
                    }
                }
            }
        },
        "internal_handlers.YieldClaimResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kadzu_sukuk-poc-be_internal_models.YieldClaim"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for accessing protected admin endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}